requirementes.txt
requests==2.*
google-ads==21.3.0
python-dotenv==1.*